[(no branch, rebasing 19.0)]: Fix state of field for geo field to 'base'
It was previously state `manual`
Usage of `manual` is for custom fields created from the interface such
as fields prefixed by `x_`.
In commit odoo/odoo@547118d manual field don't have an xmlid anymore.

Thus it had the side effect to fail when loading definition of vector layers
from an XML file as an XMLID is required to reference the field.
The XMLID would not be found.

Fixes #206